Fix scanf checks, double coefficients, and implement interval analysis
Agent-Logs-Url: https://github.com/LuizSSampaio/tp1/sessions/ff119e99-ff80-4d96-b612-18e65384637c

+++ aluno.h
@@ -1,7 +1,7 @@
 #pragma once
 
 typedef struct {
-    int coeficiente;
+    double coeficiente;
     int expoente;
 } Termo;
